+++ Assets/Scripts/Network/Player.cs
@@ -14,7 +14,7 @@ namespace Network {
         /// <value>Reference to the represented prefab</value>
         internal static GameObject Prefab { get; set; }
 
-        public int ID { get; set; }
+        public ushort ID { get; set; }
         public ushort clientID { get; set; }
         public Vector3 Position { get; set; }
 
@@ -23,7 +23,7 @@ namespace Network {
         /// </summary>
         public Player() { }
 
-        public Player(int id, ushort clientID, Vector3 position) {
+        public Player(ushort id, ushort clientID, Vector3 position) {
             this.ID = id;
             this.clientID = clientID;
             this.Position = position;
@@ -40,7 +40,7 @@ namespace Network {
         /// <param name="e">DarkRift deserialization event</param>
         public void Deserialize(DeserializeEvent e) {
             DarkRiftReader reader = e.Reader;
-            this.ID = reader.ReadInt32();
+            this.ID = reader.ReadUInt16();
             this.Position = new Vector3(reader.ReadSingle(), reader.ReadSingle(), reader.ReadSingle());
         }
